Refactor and make Pagination reusable by moving to its own component.
